Authentication-impl (#2)
* user search params instead of plain id

* Add JWT package

* * Add auth control and service
* Change config structure

* implment auth guard

* set user email as unique col

* Fix user password get update while not suppose to updating

* Add exception filter

* Add request interface

* Add auth to user

* remove debug message

+++ src/app/user/user.controller.ts
@@ -7,12 +7,23 @@ import {
 	Param,
 	Delete,
 	NotFoundException,
+	UseGuards,
+	Request,
+	UnauthorizedException,
+	UseFilters,
 } from "@nestjs/common";
 import { UserService } from "./user.service";
 import { CreateUserDto, UpdateUserDto, UserResponseDTO } from "./dto";
 import { IsInt } from "class-validator";
-import { ApiNotFoundResponse, ApiProperty, ApiTags } from "@nestjs/swagger";
+import {
+	ApiNotFoundResponse,
+	ApiProperty,
+	ApiTags,
+	ApiUnauthorizedResponse,
+} from "@nestjs/swagger";
 import { Type } from "class-transformer";
+import { AuthGuard, RequestWithUserAuthInfo } from "../auth/auth.guard";
+import { TypeOrmFilter } from "src/exceptionFilters";
 
 export class NumericIdParams {
 	@Type(() => Number)
@@ -26,6 +37,7 @@ export class NumericIdParams {
 export class UserController {
 	constructor(private readonly userService: UserService) {}
 
+	@UseFilters(TypeOrmFilter)
 	@Post()
 	async create(
 		@Body() createUserDto: CreateUserDto,
@@ -47,7 +59,9 @@ export class UserController {
 	@Get(":id")
 	@ApiNotFoundResponse({ description: "User not found" })
 	async findOne(@Param() param: NumericIdParams): Promise<UserResponseDTO> {
-		const reuslt = await this.userService.findOne(param.id);
+		const reuslt = await this.userService.findOne({
+			id: param.id,
+		});
 		if (reuslt) {
 			delete reuslt.encryptedPassword;
 			return reuslt;
@@ -58,10 +72,20 @@ export class UserController {
 
 	@Patch(":id")
 	@ApiNotFoundResponse({ description: "User not found" })
+	@ApiUnauthorizedResponse({
+		description: "You are not authorized to update this user",
+	})
+	@UseGuards(AuthGuard)
 	async update(
+		@Request() request: RequestWithUserAuthInfo,
 		@Param() param: NumericIdParams,
 		@Body() updateUserDto: UpdateUserDto,
 	) {
+		if (request.user.sub !== param.id) {
+			throw new UnauthorizedException(
+				"You are not authorized to update this user",
+			);
+		}
 		const reuslt = await this.userService.update(param.id, updateUserDto);
 		if (reuslt.affected > 0) {
 			return;
@@ -71,8 +95,20 @@ export class UserController {
 	}
 
 	@Delete(":id")
+	@ApiUnauthorizedResponse({
+		description: "You are not authorized to delete this user",
+	})
 	@ApiNotFoundResponse({ description: "User not found" })
-	async remove(@Param() param: NumericIdParams) {
+	@UseGuards(AuthGuard)
+	async remove(
+		@Request() request: RequestWithUserAuthInfo,
+		@Param() param: NumericIdParams,
+	) {
+		if (request.user.sub !== param.id) {
+			throw new UnauthorizedException(
+				"You are not authorized to delete this user",
+			);
+		}
 		const result = await this.userService.remove(param.id);
 		if (result.affected > 0) {
 			return;
